requirements.txt
fastapi
uvicorn
pandas
numpy
streamlit
plotly
requests
openai
python-dotenv
google-generativeai
